Update License
Changes from "The Unlicense" to "0-Clause BSD".
For all practical purposes, this is the same thing except that this license is recognized by OSI and applicable even in countries where public domain is not recognized. It is also compatible with jurisdiction that do not recognize copyright as being applicable to software.